Updated logging and dry run option for clarity (#17)
* Added more logging and dryrun option to clarity

To better understand why clarity is not filling all sensors I added
more diagnostic logging. I also added a dryrun option (DRYRUN) so that
you can run the code and do everthing but the actual put methods.

* More improvements on logging

* Added SOURCEID and FORCE options

SOURCEID will limit to the specific source_id you are looking for and
FORCE will force the addition of the station file even when it matches

* Reverted the max-out value to 75

+++ fetcher/lib/providers.js
@@ -1,7 +1,13 @@
 const fs = require('fs');
 const path = require('path');
 const AWS = require('aws-sdk');
-const { VERBOSE, gzip, unzip } = require('./utils');
+const {
+    VERBOSE,
+    DRYRUN,
+    gzip,
+    unzip,
+    prettyPrintStation,
+} = require('./utils');
 
 const s3 = new AWS.S3({
     maxRetries: 10
@@ -63,26 +69,32 @@ class Providers {
         try {
             const resp = await s3.getObject({ Bucket, Key }).promise();
             const currentData = (await unzip(resp.Body)).toString('utf-8');
-            if (currentData === newData) {
-                if (VERBOSE) console.log(`skip - station: ${providerStation}`);
+            if (currentData === newData && !process.env.FORCE) {
+                if (VERBOSE) console.log(`station has not changed - station: ${providerStation}`);
                 return;
             }
-            if (VERBOSE) console.log(
-                `Update ${providerStation}\n  from:\n    ${currentData}\n  to:\n    ${newData}`
-            );
+            if (VERBOSE) {
+                console.log(`-------------------------\nUpdate ${providerStation}\n----------------------> to:`);
+                prettyPrintStation(newData);
+                console.log(`-----------------> from`);
+                prettyPrintStation(currentData);
+            }
         } catch (err) {
             if (err.statusCode !== 404) throw err;
         }
 
         const compressedString = await gzip(newData);
-        await s3.putObject({
-            Bucket,
-            Key,
-            Body: compressedString,
-            ContentType: 'application/json',
-            ContentEncoding: 'gzip'
-        }).promise();
-        if (VERBOSE) console.log(`ok - station: ${providerStation}`);
+        if(!DRYRUN) {
+            if(VERBOSE) console.debug(`Saving station to ${Bucket}/${Key}`);
+            await s3.putObject({
+                Bucket,
+                Key,
+                Body: compressedString,
+                ContentType: 'application/json',
+                ContentEncoding: 'gzip'
+            }).promise();
+        }
+        if (VERBOSE) console.log(`finished station: ${providerStation}\n------------------------`);
     }
 
     /**
@@ -100,6 +112,11 @@ class Providers {
         const filename = id || `${Math.floor(Date.now() / 1000)}-${Math.random().toString(36).substring(8)}`;
         const Key = `${process.env.STACK}/measures/${provider}/${filename}.csv.gz`;
         const compressedString = await gzip(measures.csv());
+        if(DRYRUN) {
+            console.log(`Would have saved ${measures.length} measurements to '${Bucket}/${Key}'`);
+            return new Promise((y,n) => y(true));
+        }
+        if(VERBOSE) console.debug(`Saving measurements to ${Bucket}/${Key}`);
         return s3.putObject({
             Bucket,
             Key,
